refactor: <ConfigStyled /> alterado para <Config />
- Alteração do nome do componente de tela de configuração

- Alteração da fonte utilizada para o padrão correto

- Alteração do arquivo "src/router.tsx" para rotear o componente corretamente
