--- src/pages/Config/style.css
@@ -1,91 +1,0 @@
-/* import styled, { css } from 'styled-components';*/
-
-.conteiner {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-}
-
-.overlay {
-  position: fixed;
-  display: block;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0,0,0,0.5);
-  z-index: 2;
-  cursor: pointer;
-}
-
-.modal {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  width: 50%;
-  height: 360px;
-  transform: translate(-50%, -50%);
-  background-color: white;
-}
-
-.title {
-  display:flex;
-  justify-content: center;
-  align-items:center;
-  width:100%;
-  height: 20%;
-}
-
-.content {
-  display:flex;
-  justify-content:center;
-  align-items:center;
-  width:100%;
-  height: 60%;
-  gap: 28px;
-}
-
-.content-img {
-  width: 80%;
-  height: 100%;
-  border: 0px solid rgb(89, 155, 209);
-}
-
-.content-data {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  border: 0px solid blue;
-  gap: 10px;
-}
-
-.content-data-group {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  width: 100%;
-}
-
-.footer {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  width:100%;
-  height: 20%;
-  gap: 05px;
-}
-
-button {
-  height: 100%;
-  border-radius: 05;
-  background-color: rgb(89, 155, 209);
-  cursor: pointer;
-  height: 48px;
-}
