--- src/pages/ConfigStyled/style.css
@@ -1,1 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Teko:wght@300&display=swap');
